saved-jobs completed

+++ job-sera/src/app/modules/user/components/saved-jobs/saved-jobs.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { Router } from '@angular/router';
 import { AuthService } from 'src/app/modules/auth/services/auth.service';
 import { JobDetails } from 'src/app/shared/Models/job.type';
+import { CustomDialogComponent } from 'src/app/shared/components/custom-dialog/custom-dialog.component';
 import { SavedJobsService } from 'src/app/shared/service/saved-jobs.service';
 
 @Component({
@@ -10,21 +13,24 @@ import { SavedJobsService } from 'src/app/shared/service/saved-jobs.service';
 })
 export class SavedJobsComponent implements OnInit {
   savedJobs: JobDetails[] = [];
-  private userId : string = '';
-  constructor(private savedJobService: SavedJobsService, private authService: AuthService) {
+  private userId: string = '';
+  constructor(private savedJobService: SavedJobsService,
+    private authService: AuthService,
+    private dialogue: MatDialog,
+    private router: Router) {
 
   }
   ngOnInit() {
     this.userId = this.authService.currentUserIdSub.getValue()
-    if(this.userId){
+    if (this.userId) {
       this.getJobById()
     }
-      
-    
-    
+
+
+
   }
 
-  getJobById(){
+  getJobById() {
     this.savedJobService.getJobByUserId(this.userId).subscribe({
       next: res => {
         this.savedJobs = res;
@@ -35,13 +41,35 @@ export class SavedJobsComponent implements OnInit {
     })
   }
 
-  onDeleteClick(jobId:string){
-    this.savedJobService.deleteSavedJob(jobId).subscribe({
-      next: res => {
-        console.log(res);
-        this.getJobById()
+  onDeleteClick(jobId: string) {
+    const dialog = this.dialogue.open(CustomDialogComponent, {
+      width: '300px',
+      enterAnimationDuration: '100ms',
+      exitAnimationDuration: '100ms',
+      data: {
+        title: "Confirm to Delete?",
+        message: "Are you sure delete this saved jobs"
       }
     })
+    dialog.afterClosed().subscribe(res => {
+      if (res) {
+        this.savedJobService.deleteSavedJob(jobId).subscribe({
+          next: res => {
+            console.log(res);
+            this.getJobById()
+          }
+        })
+      }
+      else {
+        console.log("Your attempt reverted");
+
+      }
+    })
+
+  }
+
+  onClickApply(id:string) {
+    this.router.navigate(['job-application', id])
   }
 
 }
